task 7/7: Scope strategy-specific mutable state (TextFieldStates, Jobs, private vars) into the strategies that use them, replace fully-qualified LpBondablePool with an import, and slim the ViewModel to state + coordination (#4862)
